requirements.txt
flask
flask-cors
streamlit
python-dotenv
requests
chromadb
google-generativeai